// lib/types.ts
export type PomodoroMode = 'work' | 'short_break' | 'long_break';
export type PomodoroTechnique = 'pomodoro' | '52_17' | 'ultradian' | 'custom';

export interface PomodoroSettings {
    work: number;
    shortBreak: number;
    longBreak: number;
    longBreakInterval: number;
}

export interface PomodoroState {
    settings: PomodoroSettings;
    mode: PomodoroMode;
    secondsLeft: number;
    isActive: boolean;
    sessionsCompleted: number;
    isLoaded: boolean;
    technique: PomodoroTechnique;
    customSettings: PomodoroSettings; // To store user's custom settings
}